+++ lib/badgeville/endpoint.rb
@@ -1,0 +1,7 @@
+module Badgeville
+  class Endpoint
+    def self.client
+      Badgeville.client
+    end
+  end
+end