+++ spec/spec_helper.rb
@@ -10,5 +10,4 @@ require 'webmock/rspec'
 Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].each {|f| require f}
 
 RSpec.configure do |config|
-  
 end